% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dzil}
\alias{dzil}
\alias{pzil}
\alias{qlp}
\alias{qzil}
\alias{rzil}
\title{The Zero Inflated Laplace}
\usage{
dzil(x, mu = 0, b = 1, p = 1)

pzil(x, mu = 0, b = 1, p = 1)

qzil(q, mu = 0, b = 1, p = 1)

rzil(n, mu = 0, b = 1, p = 1)

qlp(q, mu, b)
}
\arguments{
\item{mu}{the mean of the non-zero density}

\item{b}{the shape parameter defining spread around mu
(must be >0).}

\item{p}{the zero inflation parameter (must be >0)}
}
\description{
Density, distribution function, quantile function and random
generation for the Zero Inflated Laplace distribution with mean equal
to 'mu', shape 'b' and zero inflation 'p'.

Distribution function \code{pzil}

Quantile function \code{qzil}

Random number generator \code{rzil}

Helper function in qzil
}
\details{
The mean 'mu' sets the central tendency of the Laplace, with "b"
defining the spread around 'mu'. The zero inflation parameter
"p", defines the point mass density above 0. The density under
zero relates to p as (1/(1+1)) while the probabilty density not
equal to zero is given by 1-(1/(1+1)). The function "rzil" is based
on the inverse cummulative function, and translates uniform
random numbers to their zil equivalents.
}
\examples{
curve(dnorm(x,mu=5),-2,10)
x<-rzil(100,mu=5)
hist(x)
}
\author{
Marco D. Visser
}
\seealso{
\code{\link{dnorm}}, \code{\link{dunif}},
or \code{\link{runif}}.
}
\concept{
probabilty density
}

